Move mixed PA graph replay helper to attention utils
