+++ vllm_ascend/attention/utils.py
@@ -4,6 +4,7 @@ from typing import Any
 
 import torch
 import torch.nn.functional as F
+import torch_npu
 from vllm.config import VllmConfig, get_current_vllm_config
 from vllm.distributed.kv_transfer import get_kv_transfer_group, has_kv_transfer_group, is_v1_kv_transfer_group
 from vllm.forward_context import ForwardContext, get_forward_context
@@ -17,6 +18,61 @@ from vllm_ascend.utils import (
     is_pd_decode_recompute_scheduler_enabled,
 )
 from vllm_ascend.worker.kvcomp_utils import KVCompMetaData
+
+
+@dataclass
+class PagedAttentionGraphParam:
+    """Mark PA params when PA and FIA share one graph replay list."""
+
+    params: tuple
+    layer_name: str | None
+
+
+def update_paged_attention_graph_param(
+    update_stream,
+    handle,
+    event,
+    param: PagedAttentionGraphParam,
+    block_table: torch.Tensor,
+    seq_lens: torch.Tensor,
+) -> None:
+    (
+        query,
+        key_cache,
+        value_cache,
+        num_kv_heads,
+        num_heads,
+        scale,
+        _captured_block_table,
+        _captured_seq_lens,
+        output,
+    ) = param.params
+    workspace = torch_npu._npu_paged_attention_get_workspace(
+        query=query,
+        key_cache=key_cache,
+        value_cache=value_cache,
+        num_kv_heads=num_kv_heads,
+        num_heads=num_heads,
+        scale_value=scale,
+        block_table=block_table,
+        context_lens=seq_lens,
+        out=output,
+    )
+    torch.npu.graph_task_update_begin(update_stream, handle)
+    torch_npu._npu_paged_attention(
+        query=query,
+        key_cache=key_cache,
+        value_cache=value_cache,
+        num_kv_heads=num_kv_heads,
+        num_heads=num_heads,
+        scale_value=scale,
+        block_table=block_table,
+        context_lens=seq_lens,
+        out=output,
+        workspace=workspace,
+    )
+    torch.npu.graph_task_update_end(update_stream)
+    event.record(update_stream)
 
 
 def cache_graph_workspace(
